[metrics] push method moved to pusher mod
